refactor(api/middlewares): errorHandler and notFound middlewares
