+++ api/src/middlewares/notFound.ts
@@ -1,0 +1,6 @@
+import {Request, Response, NextFunction} from 'express'
+
+export const notFound = (req: Request, res: Response, next: NextFunction) => {
+ // eslint-disable-line no-unused-vars
+  res.status(404).send(`Invalid Endpoint`)
+};
